Many improvements
