+++ package.json
@@ -47,12 +47,13 @@
     "@babel/compat-data": "~7.8.6",
     "@types/jasmine": "~2.8.8",
     "@types/jasminewd2": "~2.0.3",
-    "@types/jquery": "^3.5.0",
+    "@types/jquery": "^3.5.1",
     "@types/node": "^12.11.1",
     "@types/ol": "^6.3.1",
     "@types/proj4": "^2.5.0",
     "@types/socket.io-client": "^1.4.33",
     "codelyzer": "^6.0.0",
+    "eslint": "^7.9.0",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~5.0.0",
     "kadaster-huisstijl": "3.0.39",
@@ -64,7 +65,6 @@
     "protractor": "~7.0.0",
     "ts-node": "~8.3.0",
     "tslint": "~6.1.0",
-    "typescript": "~3.9.6",
-    "eslint": "^7.9.0"
+    "typescript": "~3.9.6"
   }
 }
