+++ src/app/components/map/map.component.scss
@@ -1,0 +1,139 @@
+@import "../../../styles";
+
+td {
+  border-top: 0px !important;
+}
+
+#category {
+  z-index: 100 !important;
+}
+
+.app-form {
+  position: absolute;
+  top: 50px;
+  height: calc(100vh - 50px);
+  z-index: 10;
+  width: 27vw;
+  min-width: 250px;
+  background-color: #ffffff;
+  box-shadow: 0 10px 10px -5px grey;
+  overflow: auto;
+  padding: 2em 1em;
+}
+
+#layers {
+  padding-right: 5em !important;
+  position: absolute;
+  right: 0;
+  width: 200px;
+  bottom: 6vh;
+  z-index: 100;
+}
+
+.feature-info {
+  position: absolute;
+  z-index: 20;
+  width: 30vw;
+  top: 20vh;
+}
+
+.backgroundmap {
+  width: 100vw;
+  height: 100vh;
+}
+
+.dropdown-menu {
+  font-size: 14px;
+  font-style: normal;
+}
+
+
+.dropdown-toggle {
+  color: white;
+}
+
+form div button {
+  box-shadow: none;
+}
+
+.open > .dropdown-toggle,
+.dropdown-toggle:hover {
+  color: rgb(19, 65, 115);
+}
+
+td {
+  border-top: 0px !important;
+}
+
+label {
+  margin-top: 5px;
+  color: #000000;
+  font-style: normal;
+  font-weight: normal;
+}
+
+input[type="text"] {
+  height: 2em;
+  padding-left: 0.4em;
+}
+
+textarea {
+  font-size: 1em;
+  padding-left: 0.4em;
+}
+
+input::placeholder, textarea::placeholder {
+  font-weight: lighter;
+  font-style: oblique;
+  font-size: 12px;
+}
+
+.invalid-input {
+  color: #FF0000;
+  font-weight: lighter;
+  font-style: italic;
+  font-size: 10px;
+}
+
+.required {
+  color: red;
+}
+
+.h-100 {
+  height: 100%;
+}
+
+#search {
+  background-color: white;
+  padding-right: 5em !important;
+  position: absolute;
+  left: 2vw;
+  width: 20em;
+  z-index: 100;
+}
+
+::ng-deep ggc-search {
+  div {
+    input {
+      height: 3em !important;
+      border-top-left-radius: 4px;
+      border-bottom-left-radius: 4px;
+      border: 1px solid grey;
+    }
+    div {
+      button:first-child {
+        height: 3em !important;
+        box-shadow: 0 0 !important;
+        border: 1px solid lightgrey !important;
+      }
+
+      button:first-child + button {
+        height: 3em !important;
+        box-shadow: 0 0 !important;
+        border-top-right-radius: 4px !important;
+        border-bottom-right-radius: 4px !important;
+        border: 1px solid lightgrey !important;
+      }
+    }
+  }
+}
