+++ src/app/form-controls/sensor-theme/sensor-theme.component.ts
@@ -1,4 +1,11 @@
-import { Component, forwardRef, OnDestroy, Input } from '@angular/core';
+import {
+  Component,
+  forwardRef,
+  OnDestroy,
+  Input,
+  AfterViewInit,
+  ChangeDetectorRef,
+} from '@angular/core';
 import { ControlValueAccessor, FormBuilder, FormControl, FormGroup, NG_VALIDATORS, NG_VALUE_ACCESSOR, Validators } from '@angular/forms';
 import { Subscription } from 'rxjs';
 
@@ -26,7 +33,7 @@ export interface SensorThemeFormValues {
     },
   ],
 })
-export class SensorThemeComponent implements ControlValueAccessor, OnDestroy {
+export class SensorThemeComponent implements ControlValueAccessor, OnDestroy, AfterViewInit {
   public form: FormGroup;
   public subscriptions: Subscription[] = [];
 
@@ -48,6 +55,7 @@ export class SensorThemeComponent implements ControlValueAccessor, OnDestroy {
   public sensorThemesList: string[];
 
   constructor(
+    private cd: ChangeDetectorRef,
     private formBuilder: FormBuilder,
   ) {
     this.form = this.formBuilder.group({
@@ -89,8 +97,6 @@ export class SensorThemeComponent implements ControlValueAccessor, OnDestroy {
       this.form.reset();
       this.form.markAsPristine();
     }
-
-    ($('.selectpicker') as any).selectpicker('refresh');
   }
 
   public registerOnTouched(fn: any) {
@@ -100,5 +106,9 @@ export class SensorThemeComponent implements ControlValueAccessor, OnDestroy {
   // communicate the inner form validation to the parent form
   public validate(_: FormControl) {
     return this.form.valid ? null : { theme: { valid: false } };
+  }
+
+  ngAfterViewInit(): void {
+    ($('.selectpicker') as any).selectpicker('refresh');
   }
 }
