+++ src/app/navbar/navbar.component.scss
@@ -19,10 +19,24 @@
     font-size: 24px;
   }
 
+  .header:hover {
+    color: #F5F5F5;
+  }
+
   h3 {
     margin-left: 15px;
     margin-right: 15px;
     color: white;
+  }
+
+  .nav > li > a:hover {
+    background-color: #eee !important;
+    border-radius: 4px 4px 0 0;
+  }
+
+  .nav > li > a.active {
+    background-color: white !important;
+    border-radius: 4px 4px 0 0;
   }
 
   a {
