+++ src/app/owner-update/owner-update.component.html
@@ -6,7 +6,7 @@
         </div>
         <div class="col-10">
             <div>
-                <h3>Update Owner information</h3>
+                <h3>Update Owner</h3>
             </div>
             <hr>
             <form [formGroup]="form" (ngSubmit)="submit()">
