+++ src/app/sensor-register/sensor-register.component.html
@@ -1,51 +1,66 @@
-<h3>Register Sensor</h3>
-<hr>
-<form [formGroup]="form" (ngSubmit)="submit()">
-    <app-sensor-type formControlName="type" [submitted]="submitted" (sensorType)="setName($event)"></app-sensor-type>
-    <div class="form-group">
-        <label> Name <span class="required">*</span>
-        </label>
-        <input class="form-control" type="text" formControlName="name" placeholder="Enter name"
-            [ngClass]="{ 'is-invalid': (f.name.dirty || f.name.touched || submitted) && f.name.errors }">
-        <div *ngIf="(f.name.dirty || f.name.touched || submitted) && f.name.errors" class="invalid-input">
-            <span class="required-text" *ngIf="f.name.errors.required">Name is required</span>
-            <span class="required-text" *ngIf="f.name.errors.minlength">Name must be at least 6 characters</span>
+<div class="container-fluid">
+    <app-navbar></app-navbar>
+    <div class="row form-header">
+        <div class="col-2">
+            <app-sidebar ></app-sidebar>
         </div>
-    </div>
-    <div class="form-group">
-        <label for="aim">Aim</label>
-        <input type="text" class="form-control" formControlName="aim" placeholder="Aim of the sensor">
-    </div>
-    <div class="form-group">
-        <label for="description">Description</label>
-        <textarea class="form-control" type="text" formControlName="description"
-            placeholder="Describe the sensor"></textarea>
-    </div>
-    <app-sensor-status formControlName="active" [submitted]="submitted"></app-sensor-status>
-    <hr>
-    <div class="form-group">
-        <label for="manufacturer"> Manufacturer <span class="required">*</span></label>
-        <input class="form-control" type="text" formControlName="manufacturer"
-            [ngClass]="{ 'is-invalid': submitted && f.manufacturer.errors }"
-            placeholder="Enter name of the manufacturer">
-        <div *ngIf="(f.manufacturer.dirty || f.manufacturer.touched || submitted) && f.manufacturer.errors"
-            class="invalid-input">
-            <span class="required-text" *ngIf="f.manufacturer.errors.required">Manufacturer is required</span>
+        <div class="col-5">
+            <div>
+                <h3>Register Sensor</h3>
+            </div>
+            <hr>
+            <form [formGroup]="form" (ngSubmit)="submit()">
+                <app-sensor-type formControlName="type" [submitted]="submitted" (sensorType)="setName($event)"></app-sensor-type>
+                <div class="form-group">
+                    <label> Name <span class="required">*</span>
+                    </label>
+                    <input class="form-control" type="text" formControlName="name" placeholder="Enter name"
+                        [ngClass]="{ 'is-invalid': (f.name.dirty || f.name.touched || submitted) && f.name.errors }">
+                    <div *ngIf="(f.name.dirty || f.name.touched || submitted) && f.name.errors" class="invalid-input">
+                        <span class="required-text" *ngIf="f.name.errors.required">Name is required</span>
+                        <span class="required-text" *ngIf="f.name.errors.minlength">Name must be at least 6 characters</span>
+                    </div>
+                </div>
+                <div class="form-group">
+                    <label for="aim">Aim</label>
+                    <input type="text" class="form-control" formControlName="aim" placeholder="Aim of the sensor">
+                </div>
+                <div class="form-group">
+                    <label for="description">Description</label>
+                    <textarea class="form-control" type="text" formControlName="description"
+                        placeholder="Describe the sensor"></textarea>
+                </div>
+                <app-sensor-status formControlName="active" [submitted]="submitted"></app-sensor-status>
+                <hr>
+                <div class="form-group">
+                    <label for="manufacturer"> Manufacturer <span class="required">*</span></label>
+                    <input class="form-control" type="text" formControlName="manufacturer"
+                        [ngClass]="{ 'is-invalid': submitted && f.manufacturer.errors }"
+                        placeholder="Enter name of the manufacturer">
+                    <div *ngIf="(f.manufacturer.dirty || f.manufacturer.touched || submitted) && f.manufacturer.errors"
+                        class="invalid-input">
+                        <span class="required-text" *ngIf="f.manufacturer.errors.required">Manufacturer is required</span>
+                    </div>
+                </div>
+                <div class="form-group">
+                    <label for="documentationUrl">Documentation URL <span class="required">*</span></label>
+                    <input class="form-control" type="url" formControlName="documentationUrl"
+                        placeholder="Enter URL of documentation"
+                        [ngClass]="{ 'is-invalid': submitted && f.documentationUrl.errors }">
+                    <div *ngIf="(f.documentationUrl.dirty || f.documentationUrl.touched || submitted) && f.documentationUrl.errors"
+                        class="invalid-input">
+                        <span class="required-text" *ngIf="f.documentationUrl.errors.required">Documentation is required</span>
+                        <span class="required-text" *ngIf="f.documentationUrl.errors.pattern">Not a valid URL</span>
+                    </div>
+                </div>
+                <app-sensor-location formControlName="location" [submitted]="submitted"></app-sensor-location>
+                <app-sensor-theme formControlName="theme" [submitted]="submitted"></app-sensor-theme>
+                <hr>
+                <button class="btn btn-primary btn-lg" type="submit"> Register </button>
+            </form>
         </div>
+        <div class="col-5" style="height: 88vh; max-height: 88vh">
+            <app-map searchBarHeight="3vh"></app-map>
+        </div>`
     </div>
-    <div class="form-group">
-        <label for="documentationUrl">Documentation URL <span class="required">*</span></label>
-        <input class="form-control" type="url" formControlName="documentationUrl"
-            placeholder="Enter URL of documentation"
-            [ngClass]="{ 'is-invalid': submitted && f.documentationUrl.errors }">
-        <div *ngIf="(f.documentationUrl.dirty || f.documentationUrl.touched || submitted) && f.documentationUrl.errors"
-            class="invalid-input">
-            <span class="required-text" *ngIf="f.documentationUrl.errors.required">Documentation is required</span>
-            <span class="required-text" *ngIf="f.documentationUrl.errors.pattern">Not a valid URL</span>
-        </div>
-    </div>
-    <app-sensor-location formControlName="location" [submitted]="submitted"></app-sensor-location>
-    <app-sensor-theme formControlName="theme" [submitted]="submitted"></app-sensor-theme>
-    <hr>
-    <button class="btn btn-primary btn-lg" type="submit"> Register </button>
-</form>
+</div>
