+++ src/app/sensor-register/sensor-register.component.scss
@@ -24,3 +24,7 @@ label {
     }
   }
 }
+
+.form-header {
+  margin-top: 72px;
+}
