+++ src/app/sensor-update/sensor-update.component.scss
@@ -18,3 +18,6 @@ button[type="submit"] {
   margin-right: 20px;
 }
 
+.form-header {
+  margin-top: 72px;
+}
