+++ src/app/sidebar/sidebar.component.html
@@ -1,17 +1,11 @@
 <nav class="sidebar">
     <div class="sidebar-sticky">
         <ul class="nav flex-column">
-            <li class="nav-item">
-                <a class="sidebar-link" routerLink="">
-                    <em class="fas fa-map-marked-alt"></em>
-                    <span class="ml-3 side-text">Map</span>
-                </a>
-            </li>
             <li class="nav-item" *ngFor="let item of navItems">
-                <a class="sidebar-link" routerLink="{{ item.route }}" routerLinkActive="active-link"
+                <a class="sidebar-link" routerLink="{{ item['route'] }}" routerLinkActive="active-link"
                    [routerLinkActiveOptions]="{exact:true}">
-                    <em class="{{ item.icon }}"></em>
-                    <span class="ml-3 side-text">{{ item.text }}</span>
+                    <em class="{{ item['icon'] }}"></em>
+                    <span class="ml-3 side-text">{{ item['text'] }}</span>
                 </a>
             </li>
         </ul>
