+++ src/app/viewer/viewer.component.scss
@@ -1,187 +1,1 @@
 @import "../../styles";
-
-.navbar {
-  position: absolute;
-  z-index: 20;
-  width: 100%;
-  height: 50px;
-  background-color: $theme-color-primary;
-  box-shadow: 0 10px 10px -5px grey;
-  padding: 0px 0px 0px 0px;
-  border-radius: 0px;
-  margin-bottom: 0;
-  justify-content: flex-start;
-
-  .header {
-    margin: 7px 15px;
-    color: white;
-    float: left;
-    font-size: 24px;
-  }
-
-  h3 {
-    margin-left: 15px;
-    margin-right: 15px;
-    color: white;
-  }
-
-  a {
-    color: $theme-color-text;
-  }
-
-  .logout {
-    a:after {
-      content: none;
-   }
-  }
-}
-
-.navbar-brand {
-  font-weight: normal;
-  font-size: 30px;
-  height: 40px;
-  color: white;
-  margin-top: 5px;
-}
-
-.logo {
-  position: absolute;
-  z-index: 10;
-  right: 0;
-  padding: 0px 0px 0px 10px;
-  margin: 30px 10px;
-}
-
-td {
-  border-top: 0px !important;
-}
-
-#category {
-  z-index: 100 !important;
-}
-
-.app-form {
-  position: absolute;
-  top: 50px;
-  height: calc(100vh - 50px);
-  z-index: 10;
-  width: 27vw;
-  min-width: 250px;
-  background-color: #ffffff;
-  box-shadow: 0 10px 10px -5px grey;
-  overflow: auto;
-  padding: 2em 1em;
-}
-
-#right-form {
-  position: absolute;
-  left: 80%;
-  width: 20%;
-  top: 20vh;
-  z-index: 100;
-}
-
-#layer-header, #search-header, #info-header {
-  float: left;
-  color: rgb(19, 65, 115);
-  background-color: white;
-  font-weight: lighter;
-  font-size: 15px;
-  text-align: right;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  padding-left: 10px;
-}
-
-#infoForm, #layersForm, #searchForm {
-  background-color: white;
-  padding-top: 10px;
-  padding-bottom: 10px;
-  padding-right: 10px;
-  margin-right: 0px;
-}
-
-#searchForm {
-  padding-right: 7em !important;
-}
-
-#layers, #search, #info {
-  margin-bottom: 10px;
-}
-
-#search-background, #layers-background, #info-background {
-  background-color: white;
-}
-
-.form-control {
-  border-radius: 4px !important;
-}
-
-.feature-info {
-  position: absolute;
-  z-index: 20;
-  width: 30vw;
-  top: 20vh;
-}
-
-.backgroundmap {
-  width: 100vw;
-  height: 100vh;
-}
-
-.dropdown-menu {
-  font-size: 14px;
-  font-style: normal;
-}
-
-
-.dropdown-toggle {
-  color: white;
-}
-
-form div button {
-  box-shadow: none;
-}
-
-.open > .dropdown-toggle,
-.dropdown-toggle:hover {
-  color: rgb(19, 65, 115);
-}
-
-td {
-  border-top: 0px !important;
-}
-
-label {
-  margin-top: 5px;
-  color: #000000;
-  font-style: normal;
-  font-weight: normal;
-}
-
-input[type="text"] {
-  height: 2em;
-  padding-left: 0.4em;
-}
-
-textarea {
-  font-size: 1em;
-  padding-left: 0.4em;
-}
-
-input::placeholder, textarea::placeholder {
-  font-weight: lighter;
-  font-style: oblique;
-  font-size: 12px;
-}
-
-.invalid-input {
-  color: #FF0000;
-  font-weight: lighter;
-  font-style: italic;
-  font-size: 10px;
-}
-
-.required {
-  color: red;
-}
